+++ src/app/user/user.component.ts
@@ -1,4 +1,4 @@
-import {Component, computed, Input, signal} from '@angular/core'; //signal umożliwia nam używanie sygnalizacji
+import {Component, computed, EventEmitter, input, Input, output, Output, signal} from '@angular/core'; //signal umożliwia nam używanie sygnalizacji
 import {DUMMY_USERS} from "../dummy-users";
 @Component({
   selector: 'app-user',
@@ -13,6 +13,14 @@ export class UserComponent {
   // Dodatkowo oznaczamy typ jako string dodatkowo ! oznacza, że może być nullable
   @Input({required: true}) public avatar!: string;
   @Input({required: true}) public name!: string;
+  //avatar = input.required<string>();
+  //name = input.required<string>();
+  //imagePath = computed(() => { return 'assets/users/'+this.avatar()})
+  @Input({required: true}) public id!: string;
+  // pozwala na emitowanie wartości na zewnątrz (dodatkowo ustalamy jaki ma być output
+  @Output() select = new EventEmitter<string>();
+  // nie wymaga tworzenia eventemittera, jednak musimy zaznaczyć jaki jest output tej funkcji output
+  select2 = output<string>();
 
   /**
    * Zwraca ścieżkę do awatara użytkownika
@@ -23,8 +31,11 @@ export class UserComponent {
 
   /**
    * Metoda wywoływana na click usera
+   * emituj identyfikator usera
+   * emit2 wykorzystuje emitter'a
    */
   public onSelectUser(){
-
+    this.select.emit(this.id);
+    this.select2.emit(this.id);
   }
 }