+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 02.01.2025
+- Tasks component
+- Working with Outputs and Emitting Data
+- Signals Input in user component
 - Input in user component (side branch 02_input)
 - Signals in user component (side branch 01_signals)
 - Improvement of user component